Changes to simpler site icon PNG system.
The complex site icon setup still had holes in it (specifically, mobile Firefox). So I'm reverting to just doing what the Tumblr site was doing which probably also has holes in it but they haven't been in the tech I use and at the moment that's feeling like about as good as I can expect.